net: sync env ethaddr to device enetaddr in eth_init()
In the previous enetaddr refactoring, the assumption with commit 56b555a644
was that the eth layer would handle the env -> device enetaddr syncing.
This was not the case as eth_initialize() is called only once and the sync
occurs there.  So make sure the eth_init() function does the env -> device
sync with every network init.


git-svn-id: svn://localhost/svn/u-boot/trunk@1987 e8ad1c53-384c-47db-b5ce-66ae1f620dea