Blackfin: update tiny board settings (voltage/default env)
From: Harald Krapfenbauer <Harald.Krapfenbauer@bluetechnix.at>

The tinyboards like to run at a little lower voltage than the default, and
they prefer to boot over the network.  For the latter, extend the common
code a little to make this easier.

Also fix the cm-bf527 env sector size while we're in here to reflect the
flash that is actually in use.


git-svn-id: svn://localhost/svn/u-boot/trunk@2047 e8ad1c53-384c-47db-b5ce-66ae1f620dea