kgdb: update mem2hex/hex2mem funcs
Convert the funcs to do the conversion inline so that we can do the copy
all at once with memcpy.  This let's us push out an weird arch-specific
issues with accessing different regions of memory to the memcpy function
like the MMRs on Blackfin systems, and it should be a bit faster.


git-svn-id: svn://localhost/svn/u-boot/trunk@2185 e8ad1c53-384c-47db-b5ce-66ae1f620dea