compiler.h: add uint typedef
Recent crc changes started using the "uint" type in headers that are used
on the build system.  This subsequently broke mingw targets as they do not
provide such a type.  So add this basic typedef to compiler.h so that we
do not have to worry about this breaking again in the future.


git-svn-id: svn://localhost/svn/u-boot/trunk@2249 e8ad1c53-384c-47db-b5ce-66ae1f620dea