[#6008] Blackfin: nand: drain the write buffer before returning
From: Andrew Caldwell <Andrew.Caldwell@analog.com>

The current Blackfin nand write function fills up the write buffer but
returns before it has had a chance to drain.  On faster systems, this
isn't a problem as the operation finishes before the ECC registers are
read, but on slower systems the ECC may be incomplete when the core tries
to read it.

So wait for the buffer to drain once we're doing writing to it.


git-svn-id: svn://localhost/svn/u-boot/trunk@2266 e8ad1c53-384c-47db-b5ce-66ae1f620dea