flash_protect: check for NULL flash info
If a flash is unable to be detected, and then someone calls flash_protect
on it (like the common code does in flash_init), the flash_protect logic
will dereference a NULL pointer.

Since flash_protect already does sanity checking on the info structs, add
a NULL pointer check in there.


git-svn-id: svn://localhost/svn/u-boot/trunk@2378 e8ad1c53-384c-47db-b5ce-66ae1f620dea