env_nand: return error when no device is found
Currently, if there is an error probing the NAND chip and the env is based
in NAND, the readenv() function will use a NULL function pointer and thus
jump to address 0.

Here I just check for a non-zero value of blocksize as that shouldn't be
zero when a valid device is found, but perhaps there is a better way for
someone familiar with the NAND internals to suggest.


git-svn-id: svn://localhost/svn/u-boot/trunk@2392 e8ad1c53-384c-47db-b5ce-66ae1f620dea