[#6262] Blackfin: propogate target cpu defines when building embedded env
Since we're no longer extracting the env from the target ELF file (since
upstream wouldn't take that change), we're back to the problem of cpu
defines not properly propogating to the env setup stage.  So the embedded
env built by the host compiler doesn't match the one that is linked into
the u-boot env.


git-svn-id: svn://localhost/svn/u-boot/trunk@2405 e8ad1c53-384c-47db-b5ce-66ae1f620dea