usb: musb: only write CLRDATATOG when appropriate
This is a change similar to what is already in the Linux driver.  We
should only program the CLRDATATOG bit when the current mode indicates
that it is needed.

usb: musb: stub out MUSB_TXCSR_MODE for Blackfin parts

The MUSB_TXCSR_MODE register setting isn't supported on Blackfin musb
parts, so stub it out to 0.  This matches Linux behavior.


git-svn-id: svn://localhost/svn/u-boot/trunk@2425 e8ad1c53-384c-47db-b5ce-66ae1f620dea