miiphy: leverage current_mii cache more
For code that uses miiphy_{read,write}, every call invokes a full look up
of the mii list.  There is already a "current_mii" cache that is used by
some code, but have the miiphy_{read,write} function use it as well.  This
does increase the code size slightly, but I think it's worth it.


git-svn-id: svn://localhost/svn/u-boot/trunk@2431 e8ad1c53-384c-47db-b5ce-66ae1f620dea