nand: allow delayed initialization
Many people like the current nand_init() behavior where it is always
initialized during boot and the flash size shown, but there are cases
where we are willing to forgo this niceness for speed/functionality.
So rather than change the default, introduce a delayed config option
people may enable.  This way the nand is only poked when someone tries
to actually use it.


git-svn-id: svn://localhost/svn/u-boot/trunk@2450 e8ad1c53-384c-47db-b5ce-66ae1f620dea