cmd_nand: document "scrub" options
The current documentation for the "scrub" option implies it takes no
options at all.  This can be annoying when you only want to scrub a
few blocks and not an entire device.  Good thing the code already
supports this though since it takes the same arguments as the "erase"
option.  Inform people!


git-svn-id: svn://localhost/svn/u-boot/trunk@2452 e8ad1c53-384c-47db-b5ce-66ae1f620dea