config.mk: unify duplicated flag setting
Multiple rules are using the expanded AFLAGS/CFLAGS settings and some are
getting so long that the rules need to be line wrapped.  So unify them in
one variable, use that variable in the rule, and then unwrap things.  This
makes the actual `make` output nicer as it doesn't have line continuations
in it anymore.


git-svn-id: svn://localhost/svn/u-boot/trunk@2525 e8ad1c53-384c-47db-b5ce-66ae1f620dea