Blackfin: drop cache tweaking when using "go"
Upstream won't take this code, so move the cache handling from the
execution of the old u-boot (which runs "go") to the new u-boot
we're jumping to.  This is by & large the most common use case of
the "go" command, so there shouldn't be an issue.

For people who want caches off when executing their own code, they
will need to use "icache off; dcache off" before doing "go".


git-svn-id: svn://localhost/svn/u-boot/trunk@2553 e8ad1c53-384c-47db-b5ce-66ae1f620dea