mmc: add generic mmc spi driver
From: Thomas Chou <thomas@wytron.com.tw>

This patch supports mmc/sd card with spi interface. It is based on
the generic mmc framework. It works with SDHC and supports multi
blocks read/write.

The crc checksum on data packet is enabled with the def,

There is a subcomamnd "mmc_spi" to setup spi bus and cs at run time.


git-svn-id: svn://localhost/svn/u-boot/trunk@2558 e8ad1c53-384c-47db-b5ce-66ae1f620dea