lib/hashtable: fix env var autocompletion
commit c81c1222427f268d29ba999c82e2477c428e7bab "Fix hash
table deletion to prevent lost entries" broke environment
variable autocompletion by accidentally inverting the
condition for 'used' in hmatch_r().

'used' is 0 if the hash table entry is not used, or -1 if deleted.
This patch makes hmatch_r actually match on valid ('used') entries,
instead of skipping them and failing to match anything.

Signed-off-by: Kim Phillips <kim.phillips@freescale.com>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>