env_flash: check for sector unaligned embedded envs
The overhaul of the env code (commit ea882baf9c17cd142c9) inadvertently
broke support for embedding the environment into a sector when it didn't
start on a sector boundary.

Since we aren't sure we want to support this behavior in the first place,
add a build check to catch this issue right away rather than at runtime.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>