Blackfin: adi boards: disable multi serial when early serial is enabled
The early serial debug logic does not work with the multi serial support
as that creates a layer between the raw serial accesses that cannot be
used early on.  So simply disable it when enabling early debug.  We want
this stuff to be as simple as possible anyways during early bring up.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>